refactor: Renamed misc step file and updated README with new step
