+++ scripts/setup_wizard/README.md
@@ -64,8 +64,9 @@ You can disable the cache for any step by changing `"cache_key": "<whatever valu
 9. Import Material Data
 10. Fix Mouth Outlines - **[Disabled by Default]**
 11. Delete Specific Objects
-12. ~~Make Character Upright~~ - **[Disabled, WIP]**
-13. Setup Head Driver
+12. Make Character Upright
+13. Set Color Management to 'Standard'
+14. Setup Head Driver
 
 
 ## Steps (Detailed Guide)
@@ -122,22 +123,25 @@ You can disable the cache for any step by changing `"cache_key": "<whatever valu
 11. Delete Specific Objects
     * This step deletes specific object(s) which is only EffectMesh at this time.
     * No selection needed.
-12. ~~Make Character Upright~~ - **[Disabled, WIP]**
+12. Make Character Upright
     * This step will set the character armature to 90 degrees on the x-axis (standing upright).
     * No selection needed.
-13. Setup Head Driver
+13. Set Color Management to 'Standard'
+    * This step will set the Color Management to Standard (normally Filmic)
+    * No selection needed.
+14. Setup Head Driver
     * This step will setup the Head Driver constraint so that face shadows work
     * No selection needed.
 
 ## Development Roadmap / Future Features
 ### Features
 - [x] Head Driver Setup
-- [ ] Make model upright if not upright (?)
+- [X] Make model upright if not upright (?)
 - [ ] Scale up x100
 - [ ] Character Ramp Type Mapping (automatically plug correct Body Ramp Type from Global Material Properties)
     - Requires knowing all characters who have a different the Body Ramp Type than the default
 - [ ] BetterFBX Support/Fix UV map imports (only one UV map is imported)
-- [ ] Color Management Filmic -> Standard
+- [X] Color Management Filmic -> Standard
 ### Refactoring
 - [ ] Refactor Material Assignment Mapping (externalize/centralize it to one locaiton)
 - [ ] Refactor Import Outline Lightmaps component
