+++ scripts/setup_wizard/import_order.py
@@ -12,7 +12,7 @@ try:
     import scripts.setup_wizard.fix_mouth_outlines
     import scripts.setup_wizard.delete_empties
     import scripts.setup_wizard.delete_specific_objects
-    import scripts.setup_wizard.make_character_upright
+    import scripts.setup_wizard.misc_final_steps
     import scripts.setup_wizard.setup_head_driver
 
     importlib.reload(scripts.setup_wizard.join_body_parts_to_body)
@@ -20,7 +20,7 @@ try:
     importlib.reload(scripts.setup_wizard.fix_mouth_outlines)
     importlib.reload(scripts.setup_wizard.delete_empties)
     importlib.reload(scripts.setup_wizard.delete_specific_objects)
-    importlib.reload(scripts.setup_wizard.make_character_upright)
+    importlib.reload(scripts.setup_wizard.misc_final_steps)
     importlib.reload(scripts.setup_wizard.setup_head_driver)
 except:
     print('Exception when trying to import required dependency scripts!')
@@ -119,9 +119,9 @@ class ComponentFunctionFactory:
         elif component_name ==  'delete_specific_objects':
             return scripts.setup_wizard.delete_specific_objects.delete_specified_objects
         elif component_name == 'make_character_upright':
-            return scripts.setup_wizard.make_character_upright.make_character_upright
+            return scripts.setup_wizard.misc_final_steps.make_character_upright
         elif component_name == 'set_color_management_to_standard':
-            return scripts.setup_wizard.make_character_upright.set_color_management_to_standard
+            return scripts.setup_wizard.misc_final_steps.set_color_management_to_standard
         elif component_name == 'setup_head_driver':
             return scripts.setup_wizard.setup_head_driver.setup_head_driver
         else:
